requirements.txt
cherrypy==7.1.0
python-Levenshtein
mecab-python3
